+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @aws-amplify/backend-cli
 
+## 0.2.0-alpha.11
+
+### Minor Changes
+
+- 56fbcc5f: Generated typescript codegen by default, and add type defaults as well
+
+### Patch Changes
+
+- Updated dependencies [56fbcc5f]
+  - @aws-amplify/model-generator@0.2.0-alpha.6
+
+## 0.2.0-alpha.10
+
+### Patch Changes
+
+- fcc7d389: Enable type checking during deployment
+- Updated dependencies [fcc7d389]
+  - @aws-amplify/sandbox@0.2.0-alpha.15
+
 ## 0.2.0-alpha.9
 
 ### Patch Changes